+++ semcore/popper/src/Popper.jsx
@@ -15,6 +15,8 @@ import useEventListener from '@semcore/utils/lib/use/useEventListener';
 import canUseDOM from '@semcore/utils/lib/canUseDOM';
 import logger from '@semcore/utils/lib/logger';
 import uniqueIDEnhancement from '@semcore/utils/lib/uniqueID';
+import { Scale } from '@semcore/animation';
+import { cssVariableEnhance } from '@semcore/utils/lib/useCssVariable';
 
 import createPopper from './createPopper';
 
@@ -57,7 +59,15 @@ class Popper extends Component {
     excludeRefs: [],
   };
 
-  static enhance = [uniqueIDEnhancement()];
+  static enhance = [
+    uniqueIDEnhancement(),
+    cssVariableEnhance({
+      variable: '--intergalactic-duration-popper',
+      fallback: '200',
+      map: Number.parseInt,
+      prop: 'duration',
+    }),
+  ];
 
   eventsInteractionMap = {
     click: {
@@ -192,6 +202,11 @@ class Popper extends Component {
         : { [name]: optionsModifiers[name] },
     }));
 
+    modifiersOptions.push({
+      name: 'computeStyles',
+      options: { gpuAcceleration: false },
+    });
+
     const modifiersMerge = [...modifiersFallback, ...modifiersOptions].concat(modifiers);
     this.popper.current = createPopper(this.triggerRef.current, this.popperRef.current, {
       placement,
@@ -324,7 +339,8 @@ class Popper extends Component {
   }
 
   getPopperProps() {
-    const { visible, disablePortal, interaction, popperZIndex, ...other } = this.asProps;
+    const { visible, disablePortal, interaction, popperZIndex, placement, duration, ...other } =
+      this.asProps;
     // @ts-ignore
     const { onKeyDown, ...interactionProps } = this.handlersFromInteraction(
       interaction,
@@ -349,6 +365,8 @@ class Popper extends Component {
       ...interactionProps,
       onKeyDown: this.bindHandlerKeyDown(onKeyDown),
       style: popperZIndex !== undefined ? { zIndex: popperZIndex } : null,
+      placement,
+      duration,
     };
   }
 
@@ -395,7 +413,7 @@ function Trigger(props) {
 }
 
 const FocusLockWrapper = React.forwardRef(function (
-  { tag, disableEnforceFocus, returnFocusRef, returnFocus, ...other },
+  { disableEnforceFocus, returnFocusRef, returnFocus, ...lockProps },
   ref,
 ) {
   const [eventLock, setEventLock] = useState(false);
@@ -432,11 +450,11 @@ const FocusLockWrapper = React.forwardRef(function (
   return (
     <FocusLock
       ref={useForkRef(popperRef, ref)}
-      as={tag}
       disabled={disableEnforceFocus || nodesLock || eventLock}
-      lockProps={other}
+      lockProps={{ ...lockProps, style: { display: 'contents', ...(lockProps.style ?? {}) } }}
       returnFocus={returnFocus}
-      {...other}
+      style={{ display: 'contents' }}
+      {...lockProps}
     />
   );
 });
@@ -452,27 +470,22 @@ function PopperPopper(props) {
     triggerRef,
     interaction,
     controlsLength,
+    duration,
   } = props;
   const ref = useRef(null);
 
   // https://github.com/facebook/react/issues/11387
   const handlerStopPropagation = useCallback((e) => e.stopPropagation(), []);
 
-  if (!visible) return null;
-
   return sstyled(styles)(
     <Portal disablePortal={disablePortal}>
       <NeighborLocation controlsLength={controlsLength}>
         <SPopper
-          render={FocusLockWrapper}
-          tag={Box}
+          render={Scale}
+          visible={visible}
+          duration={[duration, duration / 2]}
           ref={ref}
-          disableEnforceFocus={disableEnforceFocus}
           shards={[triggerRef]}
-          returnFocus={interaction === 'click'}
-          returnFocusRef={triggerRef}
-          autoFocus={false}
-          tabIndex={0}
           onClick={handlerStopPropagation}
           onContextMenu={handlerStopPropagation}
           onDoubleClick={handlerStopPropagation}
@@ -500,9 +513,17 @@ function PopperPopper(props) {
           onReset={handlerStopPropagation}
           onSubmit={handlerStopPropagation}
         >
-          <PortalProvider value={ref}>
-            <Children />
-          </PortalProvider>
+          <FocusLockWrapper
+            disableEnforceFocus={disableEnforceFocus}
+            returnFocus={interaction === 'click'}
+            returnFocusRef={triggerRef}
+            autoFocus={false}
+            tabIndex={0}
+          >
+            <PortalProvider value={ref}>
+              <Children />
+            </PortalProvider>
+          </FocusLockWrapper>
         </SPopper>
       </NeighborLocation>
     </Portal>,