+++ semcore/radio/CHANGELOG.md
@@ -2,9 +2,13 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
-## [5.3.8] - 2023-02-09
+## [5.4.0] - 2023-02-20
 
-## [5.3.7] - 2023-01-20
+### Changed
+
+- Animation duration now might be controlled with design tokens.
+
+## [5.3.8] - 2023-02-09
 
 ### Changed